requirements.txt
os
sqlite3
opencv-python
telebot
shutil
requests
zipfile
pillow
platform